+++ MLEM.Ui/Style/UiStyle.cs
@@ -31,7 +31,7 @@ namespace MLEM.Ui.Style {
         /// <summary>
         /// The color that the <see cref="Button"/> element renders with when it is moused over (<see cref="Element.IsMouseOver"/>)
         /// </summary>
-        public Color ButtonHoveredColor;
+        public Color ButtonHoveredColor = Color.LightGray;
         /// <summary>
         /// The texture that the <see cref="Button"/> element uses when it <see cref="Button.IsDisabled"/>
         /// </summary>
@@ -39,11 +39,23 @@ namespace MLEM.Ui.Style {
         /// <summary>
         /// The color that the <see cref="Button"/> element uses when it <see cref="Button.IsDisabled"/>
         /// </summary>
-        public Color ButtonDisabledColor;
+        public Color ButtonDisabledColor = Color.Gray;
         /// <summary>
         /// The texture that the <see cref="Panel"/> element uses
         /// </summary>
         public NinePatch PanelTexture;
+        /// <summary>
+        /// The <see cref="Element.ChildPadding"/> to apply to a <see cref="Panel"/> by default
+        /// </summary>
+        public Padding PanelChildPadding = new Vector2(5);
+        /// <summary>
+        /// The amount that a <see cref="Panel"/>'s scrollable area is moved per single movement of the scroll wheel
+        /// </summary>
+        public float PanelStepPerScroll = 10;
+        /// <summary>
+        /// The size of the scroller of a <see cref="Panel"/>'s scroll bar
+        /// </summary>
+        public Vector2 PanelScrollerSize = new Vector2(5, 10);
         /// <summary>
         /// The texture that the <see cref="TextField"/> element uses
         /// </summary>
@@ -55,7 +67,15 @@ namespace MLEM.Ui.Style {
         /// <summary>
         /// The color that the <see cref="TextField"/> renders with when it is moused over (<see cref="Element.IsMouseOver"/>)
         /// </summary>
-        public Color TextFieldHoveredColor;
+        public Color TextFieldHoveredColor = Color.LightGray;
+        /// <summary>
+        /// The x position that a <see cref="TextField"/>'s text should start rendering at, based on the x position of the text field
+        /// </summary>
+        public float TextFieldTextOffsetX = 4;
+        /// <summary>
+        /// The width that a <see cref="TextField"/>'s caret should render with
+        /// </summary>
+        public float TextFieldCaretWidth = 0.5F;
         /// <summary>
         /// The background texture that the <see cref="ScrollBar"/> element uses
         /// </summary>
@@ -64,6 +84,14 @@ namespace MLEM.Ui.Style {
         /// The texture that the scroll indicator of the <see cref="ScrollBar"/> element uses
         /// </summary>
         public NinePatch ScrollBarScrollerTexture;
+        /// <summary>
+        /// Whether or not a <see cref="ScrollBar"/> should use smooth scrolling
+        /// </summary>
+        public bool ScrollBarSmoothScrolling;
+        /// <summary>
+        /// The factor with which a <see cref="ScrollBar"/>'s smooth scrolling happens
+        /// </summary>
+        public float ScrollBarSmoothScrollFactor = 0.75F;
         /// <summary>
         /// The texture that the <see cref="Checkbox"/> element uses
         /// </summary>
@@ -75,11 +103,15 @@ namespace MLEM.Ui.Style {
         /// <summary>
         /// The color that the <see cref="Checkbox"/> element renders with when it is moused over (<see cref="Element.IsMouseOver"/>)
         /// </summary>
-        public Color CheckboxHoveredColor;
+        public Color CheckboxHoveredColor = Color.LightGray;
         /// <summary>
         /// The texture that the <see cref="Checkbox"/> element renders on top of its regular texture when it is <see cref="Checkbox.Checked"/>
         /// </summary>
         public TextureRegion CheckboxCheckmark;
+        /// <summary>
+        /// The width of the space between a <see cref="Checkbox"/> and its <see cref="Checkbox.Label"/>
+        /// </summary>
+        public float CheckboxTextOffsetX = 2;
         /// <summary>
         /// The texture that the <see cref="RadioButton"/> element uses
         /// </summary>
@@ -91,7 +123,7 @@ namespace MLEM.Ui.Style {
         /// <summary>
         /// The color that the <see cref="RadioButton"/> element renders with when it is moused over (<see cref="Element.IsMouseOver"/>)
         /// </summary>
-        public Color RadioHoveredColor;
+        public Color RadioHoveredColor = Color.LightGray;
         /// <summary>
         /// The texture that the <see cref="RadioButton"/> renders on top of its regular texture when it is <see cref="Checkbox.Checked"/>
         /// </summary>
@@ -103,7 +135,7 @@ namespace MLEM.Ui.Style {
         /// <summary>
         /// The offset of the <see cref="Tooltip"/> element's top left corner from the mouse position
         /// </summary>
-        public Vector2 TooltipOffset;
+        public Vector2 TooltipOffset = new Vector2(8, 16);
         /// <summary>
         /// The color that the text of a <see cref="Tooltip"/> should have
         /// </summary>
@@ -113,17 +145,25 @@ namespace MLEM.Ui.Style {
         /// </summary>
         public TimeSpan TooltipDelay = TimeSpan.Zero;
         /// <summary>
+        /// The width of a <see cref="Tooltip"/>'s default text <see cref="Paragraph"/>
+        /// </summary>
+        public float TooltipTextWidth = 50;
+        /// <summary>
+        /// The <see cref="Element.ChildPadding"/> to apply to a <see cref="Tooltip"/> by default
+        /// </summary>
+        public Padding TooltipChildPadding = new Vector2(2);
+        /// <summary>
         /// The texture that the <see cref="ProgressBar"/> element uses for its background
         /// </summary>
         public NinePatch ProgressBarTexture;
         /// <summary>
         /// The color that the <see cref="ProgressBar"/> element renders with
         /// </summary>
-        public Color ProgressBarColor;
+        public Color ProgressBarColor = Color.White;
         /// <summary>
         /// The padding that the <see cref="ProgressBar"/> uses for its progress texture (<see cref="ProgressBarProgressTexture"/>)
         /// </summary>
-        public Vector2 ProgressBarProgressPadding;
+        public Vector2 ProgressBarProgressPadding = new Vector2(1);
         /// <summary>
         /// The texture that the <see cref="ProgressBar"/> uses for displaying its progress
         /// </summary>
@@ -131,7 +171,7 @@ namespace MLEM.Ui.Style {
         /// <summary>
         /// The color that the <see cref="ProgressBar"/> renders its progress texture with
         /// </summary>
-        public Color ProgressBarProgressColor;
+        public Color ProgressBarProgressColor = Color.Red;
         /// <summary>
         /// The font that <see cref="Paragraph"/> and other elements should use for rendering.
         /// Note that, to specify a bold and italic font for <see cref="TextFormatter"/>, you should use <see cref="GenericFont.Bold"/> and <see cref="GenericFont.Italic"/>.